Unify NodeExt.kt & ParentNodeExt.kt
