+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/NodeExt.kt
@@ -1,3 +1,17 @@
 package com.bumble.appyx.navigation.node
 
+import com.bumble.appyx.navigation.children.nodeOrNull
+
 fun <T : Node<*>> T.build(): T = also { it.onBuilt() }
+
+fun Node<*>.children(): List<Node<*>> {
+    return children.value.values.mapNotNull { it.nodeOrNull }
+}
+
+inline fun <reified N : Node<*>> Node<*>.childrenOfType(): List<N> {
+    return children.value.values.mapNotNull { it.nodeOrNull as? N }
+}
+
+inline fun <reified N : Node<*>> Node<*>.firstChildOfType(): N? {
+    return children.value.values.firstOrNull { it.nodeOrNull is N }?.nodeOrNull as N?
+}
